scenario 4 - TTD for error page handling

+++ mvc-db-unit-integration-test/src/main/java/com/herbivore/springmvc/controller/GradebookController.java
@@ -45,6 +45,10 @@ public class GradebookController {
 
 	@PostMapping("/delete/student/{id}")
 	public String deleteStudent(@PathVariable int id) {
+		if (!studentAndGradeService.checkIfStudentIsNull(id)) {
+			return "error";
+		}
+
 		studentAndGradeService.deleteStudent(id);
 		return "redirect:/";
 	}
